# till-web/openapi.json
{
  "components" : {
    "schemas" : {
      "AdjustRequest" : {
        "properties" : {
          "delta" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "delta" ],
        "type" : "object"
      },
      "Committed" : {
        "properties" : {
          "committedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          }
        },
        "required" : [ "committedAt", "id" ],
        "type" : "object"
      },
      "Line" : {
        "properties" : {
          "quantity" : {
            "exclusiveMinimum" : 0,
            "format" : "int64",
            "type" : "integer"
          },
          "sku" : {
            "type" : "string"
          }
        },
        "required" : [ "quantity", "sku" ],
        "type" : "object"
      },
      "OutboxEntry" : {
        "properties" : {
          "dedupeKey" : {
            "type" : "string"
          },
          "payload" : {
            "type" : "string"
          },
          "recordedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "sequence" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "dedupeKey", "payload", "recordedAt", "sequence" ],
        "type" : "object"
      },
      "OutboxPage" : {
        "properties" : {
          "backlog" : {
            "format" : "int64",
            "type" : "integer"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/OutboxEntry"
            },
            "type" : "array"
          }
        },
        "required" : [ "backlog", "items" ],
        "type" : "object"
      },
      "Released" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "releasedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "id", "releasedAt" ],
        "type" : "object"
      },
      "Reservation" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "effectiveState" : {
            "enum" : [ "HELD", "COMMITTED", "RELEASED", "EXPIRED" ],
            "type" : "string"
          },
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "lines" : {
            "items" : {
              "$ref" : "#/components/schemas/Line"
            },
            "type" : "array"
          },
          "state" : {
            "enum" : [ "HELD", "COMMITTED", "RELEASED", "EXPIRED" ],
            "type" : "string"
          }
        },
        "required" : [ "createdAt", "effectiveState", "expiresAt", "id", "lines", "state" ],
        "type" : "object"
      },
      "ReservationPage" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/Reservation"
            },
            "type" : "array"
          }
        },
        "required" : [ "items" ],
        "type" : "object"
      },
      "ReserveRequest" : {
        "properties" : {
          "lines" : {
            "items" : {
              "$ref" : "#/components/schemas/Line"
            },
            "minItems" : 1,
            "type" : "array"
          },
          "ttlSeconds" : {
            "description" : "How long the hold lasts. Defaults to till.default-ttl.",
            "example" : 900,
            "exclusiveMinimum" : 0,
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "lines" ],
        "type" : "object"
      },
      "Reserved" : {
        "properties" : {
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "lines" : {
            "items" : {
              "$ref" : "#/components/schemas/Line"
            },
            "type" : "array"
          }
        },
        "required" : [ "expiresAt", "id", "lines" ],
        "type" : "object"
      },
      "Stock" : {
        "properties" : {
          "available" : {
            "format" : "int64",
            "type" : "integer"
          },
          "onHand" : {
            "format" : "int64",
            "type" : "integer"
          },
          "reserved" : {
            "format" : "int64",
            "type" : "integer"
          },
          "sku" : {
            "type" : "string"
          }
        },
        "required" : [ "available", "onHand", "reserved", "sku" ],
        "type" : "object"
      },
      "StockPage" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/Stock"
            },
            "type" : "array"
          },
          "nextAfter" : {
            "description" : "Absent on the last page.",
            "type" : "string"
          }
        },
        "required" : [ "items" ],
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearer" : {
        "description" : "till.auth.client-token for everything, till.auth.admin-token for changing stock levels",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "An oversell-proof inventory reservation service. Every mutating endpoint requires an Idempotency-Key header; sending the same key twice returns the first answer and changes nothing.",
    "license" : {
      "name" : "MIT",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "title" : "till",
    "version" : "0.1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/v1/outbox" : {
      "get" : {
        "operationId" : "listOutbox",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OutboxPage"
                }
              }
            },
            "description" : "the backlog and its oldest entries"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OutboxPage"
                }
              }
            },
            "description" : "this needs the admin token"
          }
        },
        "summary" : "The unpublished tail of the outbox",
        "tags" : [ "outbox-controller" ]
      }
    },
    "/v1/reservations" : {
      "get" : {
        "operationId" : "listReservations",
        "parameters" : [ {
          "in" : "query",
          "name" : "state",
          "required" : false,
          "schema" : {
            "enum" : [ "HELD", "COMMITTED", "RELEASED", "EXPIRED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 100,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationPage"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "List reservations, newest first",
        "tags" : [ "reservation-controller" ]
      },
      "post" : {
        "operationId" : "reserve",
        "parameters" : [ {
          "description" : "Retrying with this key returns the first answer and changes nothing",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReserveRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Reserved"
                }
              }
            },
            "description" : "the hold was taken, or already existed under this key"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Reserved"
                }
              }
            },
            "description" : "a SKU has never been stocked"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Reserved"
                }
              }
            },
            "description" : "not enough available stock; the body lists every shortfall"
          },
          "422" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Reserved"
                }
              }
            },
            "description" : "this key was used for a different request"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Reserved"
                }
              }
            },
            "description" : "too much contention; retry"
          }
        },
        "summary" : "Take a hold on stock",
        "tags" : [ "reservation-controller" ]
      }
    },
    "/v1/reservations/{id}" : {
      "get" : {
        "operationId" : "getReservation",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Reservation"
                }
              }
            },
            "description" : "the reservation"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Reservation"
                }
              }
            },
            "description" : "no such reservation"
          }
        },
        "summary" : "Look a hold up",
        "tags" : [ "reservation-controller" ]
      }
    },
    "/v1/reservations/{id}/commit" : {
      "post" : {
        "operationId" : "commitReservation",
        "parameters" : [ {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Committed"
                }
              }
            },
            "description" : "on-hand and reserved both went down"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Committed"
                }
              }
            },
            "description" : "no such reservation"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Committed"
                }
              }
            },
            "description" : "already committed, or already released"
          },
          "410" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Committed"
                }
              }
            },
            "description" : "the hold ran out of time"
          }
        },
        "summary" : "Turn a hold into a sale",
        "tags" : [ "reservation-controller" ]
      }
    },
    "/v1/reservations/{id}/release" : {
      "post" : {
        "operationId" : "releaseReservation",
        "parameters" : [ {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Released"
                }
              }
            },
            "description" : "the hold is gone, whether it went now or earlier"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Released"
                }
              }
            },
            "description" : "no such reservation"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Released"
                }
              }
            },
            "description" : "it was committed and cannot be released"
          }
        },
        "summary" : "Give a hold back",
        "tags" : [ "reservation-controller" ]
      }
    },
    "/v1/stock" : {
      "get" : {
        "operationId" : "listStock",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 100,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "after",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StockPage"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "List stock levels",
        "tags" : [ "stock-controller" ]
      }
    },
    "/v1/stock/{sku}" : {
      "get" : {
        "operationId" : "getStock",
        "parameters" : [ {
          "in" : "path",
          "name" : "sku",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Stock"
                }
              }
            },
            "description" : "the level"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Stock"
                }
              }
            },
            "description" : "this SKU has never been stocked"
          }
        },
        "summary" : "Read a stock level",
        "tags" : [ "stock-controller" ]
      }
    },
    "/v1/stock/{sku}/adjust" : {
      "post" : {
        "operationId" : "adjustStock",
        "parameters" : [ {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "sku",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AdjustRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Stock"
                }
              }
            },
            "description" : "the new level"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Stock"
                }
              }
            },
            "description" : "this needs the admin token"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Stock"
                }
              }
            },
            "description" : "removing stock from a SKU that has no row"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Stock"
                }
              }
            },
            "description" : "this would take on-hand below what is reserved"
          }
        },
        "summary" : "Change on-hand stock directly",
        "tags" : [ "stock-controller" ]
      }
    }
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "description" : "this service",
    "url" : "/"
  } ]
}